requirements.txt
fhir.resources>=7.1.0
faker>=20.0.0
python-dateutil>=2.8.2
openai>=1.0.0
python-dotenv>=1.0.0